Create top menu bar from scratch
