+++ src/app/dashboard/components/dashboardLayout/dashboardLayout.component.html
@@ -1,11 +1,3 @@
-<div class="wrapper">
-<app-sidebar></app-sidebar>
-  <div class="main">
-    <app-topbar></app-topbar>
-    <main class="content">
-      <div class="container-fluid p-0">
-        <router-outlet></router-outlet>
-      </div>
-    </main>
-  </div>
-</div>
+<header>
+  <app-topNav></app-topNav>
+</header>
