--- src/app/dashboard/components/sidebar/sidebar.component.html
@@ -1,28 +1,0 @@
-<nav id="sidebar" class="sidebar">
-  <div class="sidebar-content">
-
-    <a class="sidebar-brand" href="index.html">
-      <span class="align-middle">Keystone</span>
-    </a>
-
-    <ul class="sidebar-nav">
-      <ng-container *ngFor="let navSection of navStructure | keyvalue">
-        <li class="sidebar-header">
-          {{ navSection.key }}
-        </li>
-        <li *ngFor="let navItem of navSection.value | keyvalue" class="sidebar-item">
-          <a class="sidebar-link" href="{{ navItem.value['href'] }}">
-            <i class="bi bi-{{ navItem.value['icon'] }} align-middle"></i>
-            <span class="align-middle">{{ navItem.key }}</span>
-          </a>
-        </li>
-      </ng-container>
-    </ul>
-
-    <!-- Version number at the bottom of the sidebar -->
-    <div class="d-flex justify-content-center p-1">
-      <span class="text-secondary">Version {{ version }}</span>
-    </div>
-
-  </div>
-</nav>
