+++ src/app/dashboard/components/topNav/topNav.component.html
@@ -1,0 +1,30 @@
+<nav class="navbar navbar-expand-lg px-3 border-bottom">
+  <div class="container-fluid">
+
+    <!-- Navbar toggler button for mobile view -->
+    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+
+    <!-- Navbar brand (logo) -->
+    <a class="navbar-brand" href="#">Keystone</a>
+
+    <!-- Navbar user dropdown -->
+    <ul class="navbar-nav">
+      <li class="nav-item dropdown">
+        <a href="#" class="d-block link-dark text-decoration-none dropdown-toggle" id="dropdownUser" data-bs-toggle="dropdown" aria-expanded="false">
+          <img src="https://github.com/mdo.png" alt="mdo" width="32" height="32" class="rounded-circle">
+        </a>
+        <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="dropdownUser">
+          <li><a class="dropdown-item" href="#">Profile</a></li>
+          <li><a class="dropdown-item" href="#">Settings</a></li>
+          <li>
+            <hr class="dropdown-divider">
+          </li>
+          <li><a class="dropdown-item" href="#">Sign Out</a></li>
+        </ul>
+      </li>
+    </ul>
+
+  </div>
+</nav>
