+++ src/app/dashboard/components/topNav/topNav.component.scss
@@ -1,0 +1,7 @@
+@media (max-width: 576px) {
+  .navbar .navbar-brand {
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+  }
+}
