+++ src/app/dashboard/components/topNav/topNav.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { RouterOutlet } from "@angular/router";
+import { KeyValuePipe, NgForOf } from "@angular/common";
+
+@Component({
+  selector: 'app-topNav',
+  standalone: true,
+  imports: [RouterOutlet, KeyValuePipe, NgForOf],
+  templateUrl: 'topNav.component.html',
+  styleUrl: 'topNav.component.scss'
+})
+export class TopNavComponent {}
