--- src/app/dashboard/components/topbar/topbar.component.html
@@ -1,5 +1,0 @@
-<nav class="navbar navbar-expand navbar-light navbar-bg">
-  <a class="sidebar-toggle">
-    <i class="hamburger align-self-center"></i>
-  </a>
-</nav>
