--- src/app/dashboard/components/topbar/topbar.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-topbar',
-  standalone: true,
-  templateUrl: 'topbar.component.html'
-})
-export class TopbarComponent {}
