+++ src/styles.scss
@@ -1,5 +1,2 @@
-// Core theme
-@import "@adminkit/core/dist/css/app.css";
-
 // Icons
 @import "bootstrap-icons/font/bootstrap-icons.min.css";
